Added interactivity and more DB actions.
The main module pynder.py now runs a simple REPL interface to perform
basi operations on the database. This is currently achieved using a
simple loop, in the future this may be improved to achieve more
flexibility.
Interaction with the DB has been enriched with primitives to add files
and repositories. When a file added is associated to a repository the
proper entry in the relation table is added.
Tests for the new DB actions have been written and performed, currently
all tests in the suite pass.
The file 'schema.sql' has been changed to allow SQLite to manage
integer primary keys with row IDs (the SQLite preferred method for
auto-increment integer IDs).