Change in DB schema and code update.
We found a flaw in the DB design where the 'path' column was in the
'file' table and not in the 'repository_file' one (where it should
have been: a file could appear in many repositories with respect to
which it has path). This commit changes the schema.sql file (for
schema creation) and updates functions in db module also updating the
corresponding tests (currently all tests pass).